+++ sdk/search/azure-search-documents/src/main/java/com/azure/search/documents/SearchAsyncClient.java
@@ -70,6 +70,8 @@ import static com.azure.core.util.serializer.TypeReference.createInstance;
  */
 @ServiceClient(builder = SearchClientBuilder.class, isAsync = true)
 public final class SearchAsyncClient {
+    private static final ClientLogger LOGGER = new ClientLogger(SearchAsyncClient.class);
+
     /**
      * Search REST API Version
      */
@@ -84,11 +86,6 @@ public final class SearchAsyncClient {
      * The name of the Azure Cognitive Search index.
      */
     private final String indexName;
-
-    /**
-     * The logger to be used
-     */
-    private final ClientLogger logger = new ClientLogger(SearchAsyncClient.class);
 
     /**
      * The underlying AutoRest client used to interact with the Azure Cognitive Search service
@@ -596,7 +593,7 @@ public final class SearchAsyncClient {
             .collect(Collectors.toList());
 
         boolean throwOnAnyError = options == null || options.throwOnAnyError();
-        return Utility.indexDocumentsWithResponse(restClient, indexActions, throwOnAnyError, context, logger);
+        return Utility.indexDocumentsWithResponse(restClient, indexActions, throwOnAnyError, context, LOGGER);
     }
 
     /**
@@ -679,7 +676,7 @@ public final class SearchAsyncClient {
                         try {
                             return new SimpleResponse<>(res, Utility.convertValue(res.getValue(), modelClass));
                         } catch (IOException ex) {
-                            throw logger.logExceptionAsError(
+                            throw LOGGER.logExceptionAsError(
                                 new RuntimeException("Failed to deserialize document.", ex));
                         }
                     }
@@ -690,7 +687,7 @@ public final class SearchAsyncClient {
                     return new SimpleResponse<>(res, doc);
                 }).map(Function.identity());
         } catch (RuntimeException ex) {
-            return monoError(logger, ex);
+            return monoError(LOGGER, ex);
         }
     }
 
@@ -746,7 +743,7 @@ public final class SearchAsyncClient {
                 .onErrorMap(MappingUtils::exceptionMapper)
                 .map(Function.identity());
         } catch (RuntimeException ex) {
-            return monoError(logger, ex);
+            return monoError(LOGGER, ex);
         }
     }